requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-WTF>=1.0
WTForms[email]>=3.0       
SQLAlchemy>=1.4
pandas>=1.3              
openpyxl>=3.0           
python-dotenv>=0.19      
pywebview
